+++ src/kedro_graphql/utils.py
@@ -3,7 +3,7 @@ import json
 import os
 import threading
 from functools import reduce
-from typing import Any, Awaitable, TypeVar
+from typing import Awaitable, TypeVar
 from .models import Pipeline
 from urllib.parse import urlparse
 from .logs.logger import logger
@@ -109,25 +109,6 @@ def parse_s3_filepath(filepath: str) -> tuple[str, str]:
     #    raise ValueError("Invalid S3 path. S3 key (object path) is missing.")
 
     return bucket_name, s3_key, filename
-
-
-def add_param_to_feed_dict(feed_dict, param_name: str, param_value: Any, add_prefix=True) -> None:
-    """Context-free version of utility found inside KedroContext._get_feed_dict method. Option to add params: prefix if desired.
-
-    Example:
-
-        >>> param_name = "a"
-        >>> param_value = {"b": 1}
-        >>> feed_dict = {}
-        >>> _add_param_to_feed_dict(feed_dict, param_name, param_value)
-        >>> assert feed_dict["params:a"] == {"b": 1}
-        >>> assert feed_dict["params:a.b"] == 1
-    """
-    key = f"params:{param_name}" if add_prefix else param_name
-    feed_dict[key] = param_value
-    if isinstance(param_value, dict):
-        for key, val in param_value.items():
-            add_param_to_feed_dict(feed_dict, f"{param_name}.{key}", val)
 
 
 def generate_unique_paths(pipeline: Pipeline, datasets: list) -> Pipeline: